+++ example/src/components/Button/Button.style.ts
@@ -14,7 +14,7 @@ export const $Button = composeClass("Button", () => ({
   paddingVertical  : 10,
   alignItems       : "center",
   color            : "white",
-  fontWeight       : "600",
+  fontWeight       : "700",
 
   [active()]: {
     backgroundColor: "#606fc7",