+++ src/component/poly/native/PolyPressable.ts
@@ -1,6 +1,7 @@
+// @ts-ignore
 import {Pressable} from "react-native";
 import {poly}      from "../PolyComponent";
 
-export const PolyPressable = poly(Pressable);
+export const PolyPressable = Pressable && poly(Pressable);
 
 export default PolyPressable;