+++ src/component/styled/native/StyledPressable.ts
@@ -2,7 +2,7 @@ import {styled, StyledComponent} from "../StyledComponent";
 import {PropsOf}                 from "../../../Utils";
 import PolyPressable             from "../../poly/native/PolyPressable";
 
-export const StyledPressable: StyledComponent<PropsOf<typeof PolyPressable>> = styled(PolyPressable, {autoFlattens: true});
+export const StyledPressable: StyledComponent<PropsOf<typeof PolyPressable>> = PolyPressable && styled(PolyPressable, {autoFlattens: true});
 
 export type StyledPressableProps = PropsOf<typeof StyledPressable>;