+++ src/component/styled/native/StyledTextInput.ts
@@ -2,7 +2,10 @@ import {styled, StyledComponent} from "../StyledComponent";
 import {PropsOf}                 from "../../../Utils";
 import {TextInput}               from "react-native";
 
-export const StyledTextInput: StyledComponent<PropsOf<typeof TextInput>> = styled(TextInput, {autoFlattens: true, needsCascade: true});
+export const StyledTextInput: StyledComponent<PropsOf<typeof TextInput>> = styled(TextInput, {
+  autoFlattens: true,
+  needsCascade: true,
+});
 
 export type StyledTextInputProps = PropsOf<typeof StyledTextInput>;