+++ src/font/Fonts.ts
@@ -1,50 +1,62 @@
-export interface FontFamilyOptions {
-  fallbacks: string[],
+import {EnumOf} from "../Utils";
+
+export interface FontFaceOptions {
+  fallbacks?: string[],
 }
 
-export type FontWeightConfig = string | {
-  eot?: string;
-  ttf?: string;
-  woff?: string;
-  woff2?: string;
-};
+export type FontLoadListenerCallback = () => void;
 
-export interface FontFamilyConfig {
-  black?: FontWeightConfig;
-  blackItalic?: FontWeightConfig;
-  bold?: FontWeightConfig;
-  boldItalic?: FontWeightConfig;
-  extraBold?: FontWeightConfig;
-  extraBoldItalic?: FontWeightConfig;
-  extraLight?: FontWeightConfig;
-  extraLightItalic?: FontWeightConfig;
-  light?: FontWeightConfig;
-  lightItalic?: FontWeightConfig;
-  medium?: FontWeightConfig;
-  mediumItalic?: FontWeightConfig;
-  regular?: FontWeightConfig;
-  regularItalic?: FontWeightConfig;
-  semiBold?: FontWeightConfig;
-  semiBoldItalic?: FontWeightConfig;
-  thin?: FontWeightConfig;
-  thinItalic?: FontWeightConfig;
-}
+export const FontFormat = {
+  EOT  : "eot",
+  TTF  : "ttf",
+  WOFF : "woff",
+  WOFF2: "woff2",
+} as const;
+export type FontFormat = EnumOf<typeof FontFormat>;
 
-export type FontFamily = {
+export type FontWeightConfig = string | Partial<Record<FontFormat, string>>;
+
+export const FontWeight = {
+  BLACK             : "black",
+  BLACK_ITALIC      : "blackItalic",
+  BOLD              : "bold",
+  BOLD_ITALIC       : "boldItalic",
+  EXTRA_BOLD        : "extraBold",
+  EXTRA_BOLD_ITALIC : "extraBoldItalic",
+  EXTRA_LIGHT       : "extraLight",
+  EXTRA_LIGHT_ITALIC: "extraLightItalic",
+  LIGHT             : "light",
+  LIGHT_ITALIC      : "lightItalic",
+  MEDIUM            : "medium",
+  MEDIUM_ITALIC     : "mediumItalic",
+  REGULAR           : "regular",
+  REGULAR_ITALIC    : "regularItalic",
+  SEMI_BOLD         : "semiBold",
+  SEMI_BOLD_ITALIC  : "semiBoldItalic",
+  THIN              : "thin",
+  THIN_ITALIC       : "thinItalic",
+} as const;
+export type FontWeight = EnumOf<typeof FontWeight>;
+
+export type FontFaceConfig = Partial<Record<FontWeight, FontWeightConfig>>;
+
+export type FontFace = {
   (): string;
   name: string;
   weight: (weight: string) => string;
-  options: FontFamilyOptions;
-} & Record<keyof FontFamilyConfig, () => string>;
+  options: FontFaceOptions;
+  config: FontFaceConfig;
+  loadAll: () => Promise<void>;
+} & Record<FontWeight, () => string>;
 
-export const FONT_FORMAT_NAMES: Record<string, string> = {
+export const FONT_FORMAT_NAMES: Record<FontFormat, string> = {
   "ttf"  : "truetype",
   "eot"  : "embedded-opentype",
   "woff" : "woff",
   "woff2": "woff2",
 };
 
-export const FONT_WEIGHT_VALUE_TO_NAME: Record<number, keyof FontFamilyConfig> = {
+export const FONT_WEIGHT_VALUE_TO_NAME: Record<number, FontWeight> = {
   [100]: "thin",
   [200]: "extraLight",
   [300]: "light",
@@ -54,43 +66,27 @@ export const FONT_WEIGHT_VALUE_TO_NAME: Record<number, keyof FontFamilyConfig> =
   [700]: "bold",
   [800]: "extraBold",
   [900]: "black",
-};
+} as const;
 
-export const FONT_WEIGHT_NAME_TO_VALUE: Record<string, number> = {
-  "thin"      : 100,
-  "extraLight": 200,
-  "light"     : 300,
-  "regular"   : 400,
-  "medium"    : 500,
-  "semiBold"  : 600,
-  "bold"      : 700,
-  "extraBold" : 800,
-  "black"     : 900,
-};
+export const FONT_WEIGHT_NAME_TO_VALUE: Record<FontWeight, number> = {
+  "thin"            : 100,
+  "thinItalic"      : 100,
+  "extraLight"      : 200,
+  "extraLightItalic": 200,
+  "light"           : 300,
+  "lightItalic"     : 300,
+  "regular"         : 400,
+  "regularItalic"   : 400,
+  "medium"          : 500,
+  "mediumItalic"    : 500,
+  "semiBold"        : 600,
+  "semiBoldItalic"  : 600,
+  "bold"            : 700,
+  "boldItalic"      : 700,
+  "extraBold"       : 800,
+  "extraBoldItalic" : 800,
+  "black"           : 900,
+  "blackItalic"     : 900,
+} as const;
 
-export const FONT_TYPES: Array<keyof FontFamilyConfig> = [
-  "black",
-  "blackItalic",
-  "bold",
-  "boldItalic",
-  "extraBold",
-  "extraBoldItalic",
-  "extraLight",
-  "extraLightItalic",
-  "light",
-  "lightItalic",
-  "medium",
-  "mediumItalic",
-  "regular",
-  "regularItalic",
-  "semiBold",
-  "semiBoldItalic",
-  "thin",
-  "thinItalic",
-];
-
-export interface FontVariant {
-  resource: string;
-  fontFamily: FontFamily;
-  weight: number;
-}
+export type FontLoadResult = Promise<void>;