+++ src/index.ts
@@ -9,7 +9,10 @@ export * from "./component";
 
 export * from "./class/StyleClass";
 
-export {createFontFamily} from "./font/FontFamily";
+export {createFontFace, setFontBackend, setFontBackendForNative} from "./font/FontFace";
+export {default as WebFontBackend}                               from "./font/backend/WebFontBackend";
+export {default as ExpoFontBackend}                              from "./font/backend/ExpoFontBackend";
+export {default as FontBackend}                                  from "./font/backend/FontBackend";
 
 export * from "./theme";
 export * from "./unit";